fix: 修复启动脚本未加载 .env 导致 GLM API Key 失效的问题
主要修复：
- 修正 socketio_server.py 中的 .env 文件路径（从 parent.parent 改为 parent.parent.parent）
- 增强 GLM_API_KEY 加载的日志验证
- 添加详细的 LLM 调用日志（实例 ID、调用次数、参数、耗时）

其他改进：
- 统一项目命名：agent → llm（重命名 4 个实现类文件）
- 修复 VAD 音频归一化问题（int16 → float32 除以 32767.0）
- 修复前端麦克风回声问题（移除 audioContext.destination 连接）
- 添加 GLM LLM 详细日志追踪（流式调用、Token 使用、错误分类）

测试验证：
- ✅ GLM API Key 正确加载
- ✅ GLMLLM 实例创建成功（不再降级到 MockLLM）
- ✅ 前端完整流程测试通过
- ✅ LLM 调用日志正常输出

+++ src/anima/config/providers/vad/silero.py
@@ -12,7 +12,7 @@ class SileroVADConfig(VADBaseConfig):
     type: Literal["silero"] = "silero"
     sample_rate: int = Field(default=16000, description="采样率")
     prob_threshold: float = Field(default=0.4, description="语音概率阈值")
-    db_threshold: int = Field(default=60, description="分贝阈值")
+    db_threshold: float = Field(default=60.0, description="分贝阈值")
     required_hits: int = Field(default=3, description="开始说话需要的连续命中次数")
     required_misses: int = Field(default=24, description="停止说话需要的连续未命中次数")
     smoothing_window: int = Field(default=5, description="平滑窗口大小")
